Create MaxLimit
I chose to add this feature because it can be useful in scenarios where the user needs to limit the count to a certain number, such as in a game where the user has a limited number of moves or in a survey where the user can only answer a certain number of questions. Overall, I believe this feature adds more flexibility to my counting app.